% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area_km2}
\alias{area_km2}
\alias{area_m2}
\alias{area_rads2}
\title{Calculate Area of H3 Cells}
\usage{
area_km2(x)

area_m2(x)

area_rads2(x)
}
\arguments{
\item{x}{for \code{h3_from_points()} an object of class \code{sfc_POINT}. For \code{h3_from_strings()} a character vector of H3 index IDs. For \code{h3_from_xy()} a numeric vector of longitudes.}
}
\value{
a numeric vector the same length as \code{x}
}
\description{
These functions calculate the area of H3 cells for a given unit.
\itemize{
\item \code{area_km2()} calculates the area in kilometers squared
\item \code{area_m2()} calculates the area in meters squared
\item \code{area_rads2()} calculates the area in radians squared
}
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
geo <- sf::st_geometry(nc)
cells <- sfc_to_cells(geo, 5) |>
  flatten_h3()

head(area_km2(cells))
head(area_m2(cells))
head(area_rads2(cells))
}
