% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_cells.R
\name{compact_cells}
\alias{compact_cells}
\title{Compact H3 Cells}
\usage{
compact_cells(x)
}
\arguments{
\item{x}{a vector of H3 indexes}
}
\description{
Reduce a set of H3 indices of the same resolution to the minimum number of H3 indices of
varrying resolution that entirely covers the input area.
}
