% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_h3}
\alias{flatten_h3}
\title{Flatten a list H3 vectors}
\usage{
flatten_h3(x)
}
\arguments{
\item{x}{}
}
\description{
Flatten a list H3 vectors
}
\examples{
h3_ids <- c("831f91fffffffff", "831fb5fffffffff", "831f94fffffffff")

flatten_h3(
  list(
    NULL,
    h3_from_strings(h3_ids),
    h3_from_strings(h3_ids[1])
  )
)
}
