% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc-to-cells.R
\name{sfc_to_cells}
\alias{sfc_to_cells}
\title{Convert sf geometry to H3 Cells}
\usage{
sfc_to_cells(x, resolution, containment = "intersect")
}
\arguments{
\item{x}{for \code{h3_from_points()} an object of class \code{sfc_POINT}. For \code{h3_from_strings()} a character vector of H3 index IDs. For \code{h3_from_xy()} a numeric vector of longitudes.}

\item{resolution}{an integer indicating the H3 cell resolution. Must be between 0 and 15 inclusive.}

\item{containment}{character. Strategy for selecting H3 cells. Must be one of
\code{"intersect"} (Default), \code{"centroid"}, \code{"boundary"}, or \code{"covers"}. See details.}
}
\value{
An H3 vector.
}
\description{
Given a vector of sf geometries (class \code{sfc}) create a list of \code{H3} vectors.
Each list element contains the vector of H3 cells that cover the geometry.
}
\details{
Note, use \code{flatten_h3()} to reduce the list to a single vector.

The selection of H3 cells is determined by the \href{https://docs.rs/h3o/0.4.0/h3o/geom/enum.ContainmentMode.html}{\strong{Containment Mode}}:
\itemize{
\item \code{"centroid"}: Returns cells whose center point (centroid) falls inside the
geometry. This is the fastest method but may leave edges of the geometry
uncovered.
\item \code{"boundary"}: Returns only cells that are completely contained within the
geometry. This often leaves the outer perimeter of the geometry uncovered.
\item \code{"intersect"}: (Default) Returns any cell that touches the geometry (including
boundary intersections). This ensures the entire geometry is covered,
but may include cells that are only partially overlapped.
\item \code{"covers"}: An extension of \code{"intersect"}. Use this when dealing with
very small geometries. While \code{"intersect"} captures cells that touch the
geometry's boundary, \code{"covers"} ensures that if a geometry is so small
that it sits entirely inside a single cell (without touching the cell's
edges), that covering cell is still returned.
}
}
\examples{
if (interactive() && rlang::is_installed("sf")) {
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
  geo <- sf::st_geometry(nc)
  cells <- sfc_to_cells(geo, 5)

  head(cells)

  plot(flatten_h3(cells))
}

}
