% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc-to-cells.R
\name{to_cells}
\alias{to_cells}
\title{Convert sf geometry to H3 Cells}
\usage{
to_cells(x, resolution)
}
\arguments{
\item{x}{for \code{h3_from_points()} an object of class \code{sfc_POINT}. For \code{h3_from_strings()} a character vector of H3 index IDs. For \code{h3_from_xy()} a numeric vector of longitudes.}

\item{resolution}{an integer indicating the H3 cell resolution. Must be between 0 and 15 inclusive.}
}
\description{
Given a vector of sf geometries (class \code{sfc}) create a list of \code{H3} vectors.
Each list element contains the vector of H3 cells that cover the geometry.
}
\details{
Note, use \code{flatten_h3()} to reduce the list to a single vector.
}
\examples{
if (interactive() && rlang::is_installed("sf")) {
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
  geo <- sf::st_geometry(nc)
  cells <- to_cells(geo, 5)

  head(cells)

  plot(flatten_h3(cells))
}

}
