% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-constructors.R
\name{h3_from_xy}
\alias{h3_from_xy}
\alias{h3_from_points}
\alias{h3_from_strings}
\title{Create H3 Index}
\usage{
h3_from_xy(x, y, resolution)

h3_from_points(x, resolution)

h3_from_strings(x)
}
\arguments{
\item{x}{for \code{h3_from_points()} an object of class \code{sfc_POINT}. For \code{h3_from_strings()} a character vector of H3 index IDs. For \code{h3_from_xy()} a numeric vector of longitudes.}

\item{y}{a numeric vector of latitudes.}

\item{resolution}{an integer indicating the H3 cell resolution. Must be between 0 and 15 inclusive.}
}
\description{
Create H3 indices from \code{sfc} objects, vectors of x and y coordinates, or H3 string IDs.
}
\examples{
h3_from_xy(-90, 120, 5)

h3_from_strings("85f29383fffffff")

if (requireNamespace("sf")) {
  # create random points
  pnts <- sf::st_cast(
    sf::st_sfc(
      sf::st_multipoint(matrix(runif(10, max = 90), ncol = 2)),
      crs = 4326
    ),
    "POINT"
  )

  # convert to H3 objects
  h3_from_points(pnts, 5)
}
}
