% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-edges.R
\name{is_nb_pairwise}
\alias{is_nb_pairwise}
\alias{is_nb_sparse}
\title{H3 index neighbors}
\usage{
is_nb_pairwise(x, y)

is_nb_sparse(x, y)
}
\arguments{
\item{x}{an \code{H3} vector.}

\item{y}{and \code{H3} vector.}
}
\description{
H3 index neighbors
}
